Pipe input and output code moved to shared
