+++ src/Shared/Frontend/Frontend.vcxproj
@@ -24,6 +24,8 @@
     <ClCompile Include="environment.cpp" />
     <ClCompile Include="game_config.cpp" />
     <ClCompile Include="key_utility.cpp" />
+    <ClCompile Include="pipe_input.cpp" />
+    <ClCompile Include="pipe_output.cpp" />
     <ClCompile Include="replayable_input.cpp" />
     <ClCompile Include="run_game.cpp" />
     <ClCompile Include="utility.cpp" />
@@ -35,6 +37,8 @@
     <ClInclude Include="environment.h" />
     <ClInclude Include="game_config.h" />
     <ClInclude Include="key_utility.h" />
+    <ClInclude Include="pipe_input.h" />
+    <ClInclude Include="pipe_output.h" />
     <ClInclude Include="replayable_input.h" />
     <ClInclude Include="run_game.h" />
     <ClInclude Include="utility.h" />
