+++ src/Shared/Frontend/utility.cpp
@@ -3,6 +3,8 @@
 #include <chrono>
 #include <thread>
 #include <cstring>
+#include <map>
+#include <pc_gfx_charmap.h>
 #include "utility.h"
 
 std::ostream& WriteShortString(std::ostream& out, const std::string& s)
@@ -38,3 +40,64 @@ void Delay(int ms)
     std::this_thread::sleep_for(std::chrono::milliseconds(ms));
 }
 
+namespace
+{
+    std::map<int, int> unix_chars = {
+        { PASSAGE,   '#' },
+        { DOOR,      '+' },
+        { FLOOR,     '.' },
+        { PLAYER,    '@' },
+        { TRAP,      '^' },
+        { STAIRS,    '%' },
+        { GOLD,      '*' },
+        { POTION,    '!' },
+        { SCROLL,    '?' },
+        { FOOD,      ':' },
+        { STICK,     '/' },
+        { ARMOR,     ']' },
+        { AMULET,    ',' },
+        { RING,      '=' },
+        { WEAPON,    ')' },
+        { VWALL,     '|' },
+        { HWALL,     '-' },
+        { ULWALL,    '-' },
+        { URWALL,    '-' },
+        { LLWALL,    '-' },
+        { LRWALL,    '-' },
+        { 204,       '|' },
+        { 185,       '|' },
+    };
+}
+
+uint32_t CharText(uint32_t ch)
+{
+    return ch & 0x0000ffff;
+}
+
+uint32_t CharColor(uint32_t ch)
+{
+    return (ch >> 24) & 0xff;
+}
+
+bool IsText(uint32_t ch)
+{
+    return (ch & 0x010000) == 0;
+}
+
+char GetUnixChar(unsigned char c)
+{
+    auto i = unix_chars.find(c);
+    if (i != unix_chars.end())
+        return i->second;
+
+    return c;
+}
+
+char GetRawCharFromData(uint32_t* data, int r, int c, int cols)
+{
+    unsigned char ch = CharText(data[r*cols + c]);
+    auto i = unix_chars.find(ch);
+    if (i != unix_chars.end())
+        ch = i->second;
+    return (ch != 0 ? ch : ' ');
+}
