+++ src/Shared/Frontend/utility.h
@@ -20,3 +20,9 @@ std::istream& ReadShortString(std::istream& in, std::string* s);
 void throw_error(const std::string &msg);
 
 void Delay(int ms);
+
+uint32_t CharText(uint32_t ch);
+uint32_t CharColor(uint32_t ch);
+bool IsText(uint32_t ch);
+char GetUnixChar(unsigned char c);
+char GetRawCharFromData(uint32_t* data, int r, int c, int cols);
